+++ src/interfaces/location.interface.ts
@@ -1,0 +1,16 @@
+export interface ILocation {
+
+    latitude: number;
+
+    longitude: number;
+
+    city: string;
+
+    state: string;
+
+    address: string;
+
+    country: string;
+
+    postalCode: string;
+}